rename ref_bundles
