+++ python/ray/data/_internal/execution/interfaces/__init__.py
@@ -2,7 +2,7 @@ from .common import NodeIdStr
 from .execution_options import ExecutionOptions, ExecutionResources
 from .executor import Executor, OutputIterator
 from .physical_operator import PhysicalOperator
-from .ref_bundles import RefBundle
+from .ref_bundle import RefBundle
 from .task_context import TaskContext
 from .transform_fn import AllToAllTransformFn, MapTransformFn
