+++ python/ray/data/_internal/execution/interfaces/transform_fn.py
@@ -1,6 +1,6 @@
 from typing import Callable, Iterable, List, Tuple
 
-from .ref_bundles import RefBundle
+from .ref_bundle import RefBundle
 from .task_context import TaskContext
 from ray.data._internal.stats import StatsDict
 from ray.data.block import Block
